+++ extras/python/sin8192_uint8.py
@@ -28,7 +28,7 @@ def generate(outfile, tablename, tablelength, samplerate):
             
             t_x = (math.cos(2*math.pi*x-math.pi)+1)/2
 
-            scaled = int16_t(math.floor(t_x*255.999))
+            scaled = int(math.floor(t_x*255.999))
 
             outstring += str(scaled) + ', '
        ## for num in range(halftable):
